pyproject.toml
[project]
name = "jira-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp",
    "requests",
    "python-dotenv"
]
